+++ packages/abstract-observation-extractor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rosen-bridge/abstract-observation-extractor
 
+## 0.2.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/rosen-extractor@11.1.1
+
 ## 0.2.2
 
 ### Patch Changes