+++ packages/observation-extractors/bitcoin-runes-observation-extractor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rosen-bridge/bitcoin-runes-observation-extractor
 
+## 1.1.1
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-observation-extractor@0.2.3
+  - @rosen-bridge/bitcoin-scanner@0.2.3
+  - @rosen-bridge/rosen-extractor@11.1.1
+
 ## 1.1.0
 
 ### Minor Changes